EZQMS-563: Moved `ActionWithAvailability` helper type and functions from `questions` to `view`
